changed compressionAlgo in config from enum to string
