+++ core/src/main/java/apoc/util/CompressionConfig.java
@@ -10,6 +10,8 @@ import org.apache.commons.compress.compressors.lz4.BlockLZ4CompressorInputStream
 import org.apache.commons.compress.compressors.lz4.BlockLZ4CompressorOutputStream;
 import org.apache.commons.compress.compressors.snappy.FramedSnappyCompressorInputStream;
 import org.apache.commons.compress.compressors.snappy.FramedSnappyCompressorOutputStream;
+import org.apache.commons.io.IOUtils;
+
 import java.io.*;
 import java.lang.reflect.Constructor;
 import java.nio.charset.Charset;
@@ -22,65 +24,16 @@ import static apoc.util.Util.convertFromListToBytes;
 
 public class CompressionConfig {
 
-    enum CompressionAlgo {
-
-        GZIP(GzipCompressorOutputStream.class, GzipCompressorInputStream.class),
-        BZIP2(BZip2CompressorOutputStream.class, BZip2CompressorInputStream.class),
-        DEFLATE(DeflateCompressorOutputStream.class, DeflateCompressorInputStream.class),
-        BLOCK_LZ4(BlockLZ4CompressorOutputStream.class, BlockLZ4CompressorInputStream.class),
-        FRAMED_SNAPPY(FramedSnappyCompressorOutputStream.class, FramedSnappyCompressorInputStream.class);
-
-        private final Class<?> compressor;
-        private final Class<?> decompressor;
-
-        CompressionAlgo(Class<?> compressor, Class<?> decompressor) {
-            this.compressor = compressor;
-            this.decompressor = decompressor;
-        }
-
-        public List<Long> compress(String string, Charset charset) throws Exception {
-            Constructor<?> constructor = compressor.getConstructor(OutputStream.class);
-            try (ByteArrayOutputStream stream = new ByteArrayOutputStream()) {
-                try (OutputStream outputStream = (OutputStream) constructor.newInstance((OutputStream) stream)) {
-                    outputStream.write(string.getBytes(charset));
-                }
-                return convertFromBytesToList(stream.toByteArray());
-            }
-        }
-
-        public String decompress(List<Long> byteArray, Charset charset) throws Exception {
-            try (ByteArrayInputStream stream = new ByteArrayInputStream(convertFromListToBytes(byteArray))) {
-                Constructor<?> constructor = decompressor.getConstructor(InputStream.class);
-                try (InputStream inputStream = (InputStream) constructor.newInstance((InputStream) stream)) {
-                    return inputReader(inputStream, charset);
-                }
-            }
-        }
-
-        private static String inputReader(InputStream inputStream, Charset charset) throws IOException {
-            try (InputStreamReader inputStreamReader = new InputStreamReader(inputStream, charset)) {
-                try (BufferedReader bufferedReader = new BufferedReader(inputStreamReader)) {
-                    StringBuilder output = new StringBuilder();
-                    String line;
-                    while ((line = bufferedReader.readLine()) != null) {
-                        output.append(line);
-                    }
-                    return output.toString();
-                }
-            }
-        }
-    }
-
-    private final CompressionAlgo compressionAlgo;
+    private final String compressionAlgo;
     private final Charset charset;
 
     public CompressionConfig(Map<String, Object> config) {
         if (config == null) config = Collections.emptyMap();
-        this.compressionAlgo = CompressionAlgo.valueOf((String) config.getOrDefault("compression", "GZIP"));
+        this.compressionAlgo = (String) config.getOrDefault("compression", "GZIP");
         this.charset = Charset.forName((String) config.getOrDefault("charset", "UTF-8"));
     }
 
-    public CompressionAlgo getCompressionAlgo() {
+    public String getCompressionAlgo() {
         return compressionAlgo;
     }
